Move k8s.io/apiserver/pkg/util/version to component-base
