+++ staging/src/k8s.io/apiserver/pkg/server/options/serving_test.go
@@ -43,10 +43,10 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/serializer"
 	"k8s.io/apimachinery/pkg/version"
 	"k8s.io/apiserver/pkg/server"
-	utilversion "k8s.io/apiserver/pkg/util/version"
 	"k8s.io/client-go/discovery"
 	restclient "k8s.io/client-go/rest"
 	cliflag "k8s.io/component-base/cli/flag"
+	utilversion "k8s.io/component-base/registry"
 	"k8s.io/klog/v2/ktesting"
 	netutils "k8s.io/utils/net"
 )
