+++ staging/src/k8s.io/component-base/registry/registry_test.go
